requirements.txt
anytree==2.4.3
bs4==0.0.1
cachecontrol[filecache]==0.12.4
coverage==4.5.1
discord.py[voice]==0.16.12
flask==0.12.2
flask-babel==0.11.2
ftfy==5.3.0
google-api-python-client==1.6.5
git+https://github.com/PennyDreadfulMTG/googler.git
inflect==0.2.5
isort==4.2.15
matplotlib==2.1.2
markdown==2.6.11
munch==2.2.0
mysqlclient==1.3.12
mypy==0.560
poeditor==1.0.5
pygithub==1.36
pygtrie==2.2
pylint==1.8.2
pystache==0.5.4
pytest==3.4.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
requests_oauthlib==0.8.0
seaborn==0.8.1
titlecase==0.12.0
untangle==1.1.1
whoosh==2.7.4
